requirements.txt
psycopg2-binary==2.9.6
fastapi==0.98.0
pydantic==1.10.9
aiohttp==3.8.4
uvicorn==0.22.0
black==23.3.0
pylint==2.17.4
httpx==0.24.1
